feat: enhance note management with user authentication and authorization checks

+++ src/note/note.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { Prisma } from '@prisma/client';
 
@@ -7,8 +11,17 @@ export class NoteService {
   constructor(private prisma: PrismaService) {}
 
   // Create a new note
+  // one way to do it
+  // async create(data: Prisma.NoteUncheckedCreateInput) {
+  //   return this.prisma.note.create({
+  //     data,
+  //   });
+  // }
+
   async create(data: Prisma.NoteCreateInput) {
-    return this.prisma.note.create({ data });
+    return this.prisma.note.create({
+      data,
+    });
   }
 
   // Get all notes
@@ -26,17 +39,40 @@ export class NoteService {
   }
 
   // Update note by ID
-  async update(id: string, data: Partial<{ title: string; content: string }>) {
-    try {
-      return await this.prisma.note.update({ where: { id }, data });
-    } catch {
-      throw new NotFoundException('Note not found');
+  async update(
+    id: string,
+    userId: string,
+    data: Partial<{ title: string; content: string }>,
+  ) {
+    // Find the note first
+    const note = await this.prisma.note.findUnique({
+      where: { id },
+    });
+
+    // Check if it exists and belongs to this user
+    if (!note || note.userId !== userId) {
+      throw new ForbiddenException('You are not allowed to edit this note');
     }
+
+    // Proceed to update
+    return this.prisma.note.update({
+      where: { id },
+      data,
+    });
   }
 
   // Delete note by ID
-  async remove(id: string) {
+  async remove(id: string, userId: string) {
     try {
+      // Find the note first
+      const note = await this.prisma.note.findUnique({
+        where: { id },
+      });
+
+      if (note && note.userId !== userId) {
+        throw new ForbiddenException('You are not allowed to delete this note');
+      }
+
       return await this.prisma.note.delete({ where: { id } });
     } catch {
       throw new NotFoundException('Note not found');
